refactor: extract globe module
Moves Globe3D (Three.js sphere with pins, trips, stars, grid),
GlobeFallback2D (Canvas2D fallback), the GLOBE_CONTINENTS polygon
data, and the Nominatim geocodeLocations() helper out of js/main.js
into the new js/globe.js module.

main.js imports those exports back and re-exports them so the test
surface is unchanged.

207 tests still pass.
